pyproject.toml
[tool.poetry]
name = "jsonschema-cli"
version = "0.1.0"
description = ""
authors = ["Eyal Mor <eyalmor94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
black = "^19.10b0"
flake8-black = "^0.2.0"
flake8-import-order = "^0.18.1"
flake8-blind-except = "^0.1.1"
flake8-builtins = "^1.5.3"
flake8-logging-format = "^0.6.0"
flake8-docstrings = "^1.5.0"
flake8-rst-docstrings = "^0.0.13"
ipython = "^7.14.0"
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120